+++ tests/proofmode.v
@@ -1480,6 +1480,60 @@ Proof.
   - iDestruct "H" as "[_ [_ #$]]".
 Qed.
 
+(* [from_and_pure_iff] and [from_sep_pure_if] instances work correctly. *)
+Lemma test_iSplit_pure_iff (φ : Prop) :
+  ⊢ ⌜φ ↔ φ⌝ : PROP.
+Proof.
+  by iSplit.
+Restart. Proof.
+  by iSplitL.
+Restart. Proof.
+  by iSplitR.
+Qed.
+
+(* Behavior of [⌜φ ↔ ψ⌝] with respect to [iSplit] is consistent with Iris-level
+ * [∗-∗] and [↔]. *)
+(* [iSplit{L,R}] fail without [BiAffine] instance. *)
+Check "test_iSplit_wand_iff".
+Lemma test_iSplit_wand_iff (P : PROP) :
+  ⊢ P ∗-∗ P.
+Proof.
+  iSplit; by auto.
+Restart.
+  (* fails without [BiAffine] instance *)
+  Fail iSplitL. Fail iSplitR.
+Abort.
+
+Lemma test_iSplit_bi_wand_iff_affine `{!BiAffine PROP} (P : PROP) :
+  ⊢ P ∗-∗ P.
+Proof.
+  iSplit; by auto.
+Restart.
+  iSplitL; by auto.
+Restart. Proof.
+  iSplitR; by auto.
+Qed.
+
+Check "test_iSplit_iff".
+Lemma test_iSplit_iff (P : PROP) :
+  ⊢ P ↔ P.
+Proof.
+  iSplit; by auto.
+Restart.
+  (* fails without [BiAffine] instance *)
+  Fail iSplitL. Fail iSplitR.
+Abort.
+
+Lemma test_iSplit_bi_iff_affine `{!BiAffine PROP} (P : PROP) :
+  ⊢ P ↔ P.
+Proof.
+  iSplit; by auto.
+Restart.
+  iSplitL; by auto.
+Restart. Proof.
+  iSplitR; by auto.
+Qed.
+
 Check "test_and_sep_affine_bi".
 Lemma test_and_sep_affine_bi `{!BiAffine PROP} P Q : □ P ∧ Q ⊢ □ P ∗ Q.
 Proof.